Delete MeterSharedState
