+++ sdk/metrics/src/main/java/io/opentelemetry/sdk/metrics/SdkLongGauge.java
@@ -15,23 +15,19 @@ import io.opentelemetry.api.metrics.ObservableLongMeasurement;
 import io.opentelemetry.context.Context;
 import io.opentelemetry.sdk.metrics.internal.descriptor.Advice;
 import io.opentelemetry.sdk.metrics.internal.descriptor.InstrumentDescriptor;
-import io.opentelemetry.sdk.metrics.internal.state.MeterProviderSharedState;
-import io.opentelemetry.sdk.metrics.internal.state.MeterSharedState;
 import io.opentelemetry.sdk.metrics.internal.state.WriteableMetricStorage;
 import java.util.List;
 import java.util.function.Consumer;
 
 final class SdkLongGauge extends AbstractInstrument implements ExtendedLongGauge {
 
-  private final MeterSharedState meterSharedState;
+  private final SdkMeter sdkMeter;
   private final WriteableMetricStorage storage;
 
   private SdkLongGauge(
-      InstrumentDescriptor descriptor,
-      MeterSharedState meterSharedState,
-      WriteableMetricStorage storage) {
+      InstrumentDescriptor descriptor, SdkMeter sdkMeter, WriteableMetricStorage storage) {
     super(descriptor);
-    this.meterSharedState = meterSharedState;
+    this.sdkMeter = sdkMeter;
     this.storage = storage;
   }
 
@@ -52,7 +48,7 @@ final class SdkLongGauge extends AbstractInstrument implements ExtendedLongGauge
 
   @Override
   public boolean isEnabled() {
-    return meterSharedState.isMeterEnabled() && storage.isEnabled();
+    return sdkMeter.isMeterEnabled() && storage.isEnabled();
   }
 
   static final class SdkLongGaugeBuilder implements ExtendedLongGaugeBuilder {
@@ -60,20 +56,14 @@ final class SdkLongGauge extends AbstractInstrument implements ExtendedLongGauge
     private final InstrumentBuilder builder;
 
     SdkLongGaugeBuilder(
-        MeterProviderSharedState meterProviderSharedState,
-        MeterSharedState sharedState,
+        SdkMeter sdkMeter,
         String name,
         String description,
         String unit,
         Advice.AdviceBuilder adviceBuilder) {
 
       builder =
-          new InstrumentBuilder(
-                  name,
-                  InstrumentType.GAUGE,
-                  InstrumentValueType.LONG,
-                  meterProviderSharedState,
-                  sharedState)
+          new InstrumentBuilder(name, InstrumentType.GAUGE, InstrumentValueType.LONG, sdkMeter)
               .setDescription(description)
               .setUnit(unit)
               .setAdviceBuilder(adviceBuilder);
