+++ sdk/metrics/src/main/java/io/opentelemetry/sdk/metrics/SdkMeter.java
@@ -5,6 +5,10 @@
 
 package io.opentelemetry.sdk.metrics;
 
+import static java.util.stream.Collectors.toMap;
+
+import io.opentelemetry.api.common.Attributes;
+import io.opentelemetry.api.internal.GuardedBy;
 import io.opentelemetry.api.metrics.BatchCallback;
 import io.opentelemetry.api.metrics.DoubleGaugeBuilder;
 import io.opentelemetry.api.metrics.DoubleHistogramBuilder;
@@ -13,20 +17,30 @@ import io.opentelemetry.api.metrics.LongUpDownCounterBuilder;
 import io.opentelemetry.api.metrics.Meter;
 import io.opentelemetry.api.metrics.MeterProvider;
 import io.opentelemetry.api.metrics.ObservableMeasurement;
+import io.opentelemetry.context.Context;
 import io.opentelemetry.sdk.common.InstrumentationScopeInfo;
 import io.opentelemetry.sdk.metrics.data.MetricData;
 import io.opentelemetry.sdk.metrics.internal.MeterConfig;
+import io.opentelemetry.sdk.metrics.internal.descriptor.InstrumentDescriptor;
 import io.opentelemetry.sdk.metrics.internal.export.RegisteredReader;
+import io.opentelemetry.sdk.metrics.internal.state.AsynchronousMetricStorage;
 import io.opentelemetry.sdk.metrics.internal.state.CallbackRegistration;
 import io.opentelemetry.sdk.metrics.internal.state.MeterProviderSharedState;
-import io.opentelemetry.sdk.metrics.internal.state.MeterSharedState;
+import io.opentelemetry.sdk.metrics.internal.state.MetricStorage;
+import io.opentelemetry.sdk.metrics.internal.state.MetricStorageRegistry;
 import io.opentelemetry.sdk.metrics.internal.state.SdkObservableMeasurement;
+import io.opentelemetry.sdk.metrics.internal.state.SynchronousMetricStorage;
+import io.opentelemetry.sdk.metrics.internal.state.WriteableMetricStorage;
+import io.opentelemetry.sdk.metrics.internal.view.RegisteredView;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
+import java.util.Map;
+import java.util.Objects;
 import java.util.Set;
+import java.util.function.Function;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 import java.util.regex.Pattern;
@@ -53,9 +67,16 @@ final class SdkMeter implements Meter {
   private static final Meter NOOP_METER = MeterProvider.noop().get("noop");
   private static final String NOOP_INSTRUMENT_NAME = "noop";
 
-  private final InstrumentationScopeInfo instrumentationScopeInfo;
+  private final Object collectLock = new Object();
+  private final Object callbackLock = new Object();
+
+  @GuardedBy("callbackLock")
+  private final List<CallbackRegistration> callbackRegistrations = new ArrayList<>();
+
   private final MeterProviderSharedState meterProviderSharedState;
-  private final MeterSharedState meterSharedState;
+  private final InstrumentationScopeInfo instrumentationScopeInfo;
+  private final Map<RegisteredReader, MetricStorageRegistry> readerStorageRegistries;
+  private final boolean meterEnabled;
 
   SdkMeter(
       MeterProviderSharedState meterProviderSharedState,
@@ -64,8 +85,10 @@ final class SdkMeter implements Meter {
       MeterConfig meterConfig) {
     this.instrumentationScopeInfo = instrumentationScopeInfo;
     this.meterProviderSharedState = meterProviderSharedState;
-    this.meterSharedState =
-        MeterSharedState.create(instrumentationScopeInfo, registeredReaders, meterConfig);
+    this.readerStorageRegistries =
+        registeredReaders.stream()
+            .collect(toMap(Function.identity(), unused -> new MetricStorageRegistry()));
+    this.meterEnabled = meterConfig.isEnabled();
   }
 
   // Visible for testing
@@ -75,41 +98,76 @@ final class SdkMeter implements Meter {
 
   /** Collect all metrics for the meter. */
   Collection<MetricData> collectAll(RegisteredReader registeredReader, long epochNanos) {
-    return meterSharedState.collectAll(registeredReader, meterProviderSharedState, epochNanos);
+    // Short circuit collection process if meter is disabled
+    if (!meterEnabled) {
+      return Collections.emptyList();
+    }
+    List<CallbackRegistration> currentRegisteredCallbacks;
+    synchronized (callbackLock) {
+      currentRegisteredCallbacks = new ArrayList<>(callbackRegistrations);
+    }
+    // Collections across all readers are sequential
+    synchronized (collectLock) {
+      for (CallbackRegistration callbackRegistration : currentRegisteredCallbacks) {
+        callbackRegistration.invokeCallback(
+            registeredReader, meterProviderSharedState.getStartEpochNanos(), epochNanos);
+      }
+
+      Collection<MetricStorage> storages =
+          Objects.requireNonNull(readerStorageRegistries.get(registeredReader)).getStorages();
+      List<MetricData> result = new ArrayList<>(storages.size());
+      for (MetricStorage storage : storages) {
+        MetricData current =
+            storage.collect(
+                meterProviderSharedState.getResource(),
+                getInstrumentationScopeInfo(),
+                meterProviderSharedState.getStartEpochNanos(),
+                epochNanos);
+        // Ignore if the metric data doesn't have any data points, for example when aggregation is
+        // Aggregation#drop()
+        if (!current.isEmpty()) {
+          result.add(current);
+        }
+      }
+      return Collections.unmodifiableList(result);
+    }
   }
 
-  /** Reset the meter, clearing all registered instruments. */
+  /** Reset the meter, clearing all registered callbacks and storages. */
   void resetForTest() {
-    this.meterSharedState.resetForTest();
+    synchronized (collectLock) {
+      synchronized (callbackLock) {
+        callbackRegistrations.clear();
+      }
+      this.readerStorageRegistries.values().forEach(MetricStorageRegistry::resetForTest);
+    }
   }
 
   @Override
   public LongCounterBuilder counterBuilder(String name) {
     return checkValidInstrumentName(name)
-        ? new SdkLongCounter.SdkLongCounterBuilder(meterProviderSharedState, meterSharedState, name)
+        ? new SdkLongCounter.SdkLongCounterBuilder(this, name)
         : NOOP_METER.counterBuilder(NOOP_INSTRUMENT_NAME);
   }
 
   @Override
   public LongUpDownCounterBuilder upDownCounterBuilder(String name) {
     return checkValidInstrumentName(name)
-        ? new SdkLongUpDownCounter.SdkLongUpDownCounterBuilder(
-            meterProviderSharedState, meterSharedState, name)
+        ? new SdkLongUpDownCounter.SdkLongUpDownCounterBuilder(this, name)
         : NOOP_METER.upDownCounterBuilder(NOOP_INSTRUMENT_NAME);
   }
 
   @Override
   public DoubleHistogramBuilder histogramBuilder(String name) {
     return checkValidInstrumentName(name)
-        ? new SdkDoubleHistogram.SdkDoubleHistogramBuilder(
-            meterProviderSharedState, meterSharedState, name)
+        ? new SdkDoubleHistogram.SdkDoubleHistogramBuilder(this, name)
         : NOOP_METER.histogramBuilder(NOOP_INSTRUMENT_NAME);
   }
 
   @Override
   public DoubleGaugeBuilder gaugeBuilder(String name) {
     return checkValidInstrumentName(name)
-        ? new SdkDoubleGauge.SdkDoubleGaugeBuilder(meterProviderSharedState, meterSharedState, name)
+        ? new SdkDoubleGauge.SdkDoubleGaugeBuilder(this, name)
         : NOOP_METER.gaugeBuilder(NOOP_INSTRUMENT_NAME);
   }
 
@@ -131,9 +189,7 @@ final class SdkMeter implements Meter {
         continue;
       }
       SdkObservableMeasurement sdkMeasurement = (SdkObservableMeasurement) measurement;
-      if (!meterSharedState
-          .getInstrumentationScopeInfo()
-          .equals(sdkMeasurement.getInstrumentationScopeInfo())) {
+      if (!instrumentationScopeInfo.equals(sdkMeasurement.getInstrumentationScopeInfo())) {
         logger.log(
             Level.WARNING,
             "batchCallback called with instruments that belong to a different Meter.");
@@ -144,8 +200,89 @@ final class SdkMeter implements Meter {
 
     CallbackRegistration callbackRegistration =
         CallbackRegistration.create(sdkMeasurements, callback);
-    meterSharedState.registerCallback(callbackRegistration);
-    return new SdkObservableInstrument(meterSharedState, callbackRegistration);
+    registerCallback(callbackRegistration);
+    return new SdkObservableInstrument(this, callbackRegistration);
+  }
+
+  /**
+   * Unregister the callback.
+   *
+   * <p>Callbacks are originally registered via {@link #registerCallback(CallbackRegistration)}.
+   */
+  void removeCallback(CallbackRegistration callbackRegistration) {
+    synchronized (callbackLock) {
+      this.callbackRegistrations.remove(callbackRegistration);
+    }
+  }
+
+  /**
+   * Register the callback.
+   *
+   * <p>The callback will be invoked once per collection until unregistered via {@link
+   * #removeCallback(CallbackRegistration)}.
+   */
+  void registerCallback(CallbackRegistration callbackRegistration) {
+    synchronized (callbackLock) {
+      callbackRegistrations.add(callbackRegistration);
+    }
+  }
+
+  /** Returns {@code true} if the {@link MeterConfig#enabled()} of the meter is {@code true}. */
+  boolean isMeterEnabled() {
+    return meterEnabled;
+  }
+
+  /** Registers new synchronous storage associated with a given instrument. */
+  WriteableMetricStorage registerSynchronousMetricStorage(InstrumentDescriptor instrument) {
+
+    List<SynchronousMetricStorage> registeredStorages = new ArrayList<>();
+    for (Map.Entry<RegisteredReader, MetricStorageRegistry> entry :
+        readerStorageRegistries.entrySet()) {
+      RegisteredReader reader = entry.getKey();
+      MetricStorageRegistry registry = entry.getValue();
+      for (RegisteredView registeredView :
+          reader.getViewRegistry().findViews(instrument, getInstrumentationScopeInfo())) {
+        if (Aggregation.drop() == registeredView.getView().getAggregation()) {
+          continue;
+        }
+        registeredStorages.add(
+            registry.register(
+                SynchronousMetricStorage.create(
+                    reader,
+                    registeredView,
+                    instrument,
+                    meterProviderSharedState.getExemplarFilter())));
+      }
+    }
+
+    if (registeredStorages.size() == 1) {
+      return registeredStorages.get(0);
+    }
+
+    return new MultiWritableMetricStorage(registeredStorages);
+  }
+
+  /** Register new asynchronous storage associated with a given instrument. */
+  SdkObservableMeasurement registerObservableMeasurement(
+      InstrumentDescriptor instrumentDescriptor) {
+    List<AsynchronousMetricStorage<?, ?>> registeredStorages = new ArrayList<>();
+    for (Map.Entry<RegisteredReader, MetricStorageRegistry> entry :
+        readerStorageRegistries.entrySet()) {
+      RegisteredReader reader = entry.getKey();
+      MetricStorageRegistry registry = entry.getValue();
+      for (RegisteredView registeredView :
+          reader.getViewRegistry().findViews(instrumentDescriptor, getInstrumentationScopeInfo())) {
+        if (Aggregation.drop() == registeredView.getView().getAggregation()) {
+          continue;
+        }
+        registeredStorages.add(
+            registry.register(
+                AsynchronousMetricStorage.create(reader, registeredView, instrumentDescriptor)));
+      }
+    }
+
+    return SdkObservableMeasurement.create(
+        instrumentationScopeInfo, instrumentDescriptor, registeredStorages);
   }
 
   @Override
@@ -169,5 +306,37 @@ final class SdkMeter implements Meter {
     }
 
     return false;
+  }
+
+  private static class MultiWritableMetricStorage implements WriteableMetricStorage {
+    private final List<? extends WriteableMetricStorage> storages;
+
+    private MultiWritableMetricStorage(List<? extends WriteableMetricStorage> storages) {
+      this.storages = storages;
+    }
+
+    @Override
+    public void recordLong(long value, Attributes attributes, Context context) {
+      for (WriteableMetricStorage storage : storages) {
+        storage.recordLong(value, attributes, context);
+      }
+    }
+
+    @Override
+    public void recordDouble(double value, Attributes attributes, Context context) {
+      for (WriteableMetricStorage storage : storages) {
+        storage.recordDouble(value, attributes, context);
+      }
+    }
+
+    @Override
+    public boolean isEnabled() {
+      for (WriteableMetricStorage storage : storages) {
+        if (storage.isEnabled()) {
+          return true;
+        }
+      }
+      return false;
+    }
   }
 }
