+++ sdk/metrics/src/main/java/io/opentelemetry/sdk/metrics/internal/state/MeterProviderSharedState.java
@@ -41,5 +41,5 @@ public abstract class MeterProviderSharedState {
   public abstract long getStartEpochNanos();
 
   /** Returns the {@link ExemplarFilter} for remembering synchronous measurements. */
-  abstract ExemplarFilter getExemplarFilter();
+  public abstract ExemplarFilter getExemplarFilter();
 }
