+++ sdk/metrics/src/test/java/io/opentelemetry/sdk/metrics/InstrumentBuilderTest.java
@@ -12,7 +12,6 @@ import io.opentelemetry.sdk.metrics.internal.MeterConfig;
 import io.opentelemetry.sdk.metrics.internal.descriptor.Advice;
 import io.opentelemetry.sdk.metrics.internal.exemplar.ExemplarFilter;
 import io.opentelemetry.sdk.metrics.internal.state.MeterProviderSharedState;
-import io.opentelemetry.sdk.metrics.internal.state.MeterSharedState;
 import io.opentelemetry.sdk.resources.Resource;
 import io.opentelemetry.sdk.testing.time.TestClock;
 import java.util.Collections;
@@ -24,18 +23,15 @@ class InstrumentBuilderTest {
       MeterProviderSharedState.create(
           TestClock.create(), Resource.getDefault(), ExemplarFilter.alwaysOff(), 0);
   static final InstrumentationScopeInfo SCOPE = InstrumentationScopeInfo.create("scope-name");
-  public static final MeterSharedState METER_SHARED_STATE =
-      MeterSharedState.create(SCOPE, Collections.emptyList(), MeterConfig.defaultConfig());
+  public static final SdkMeter SDK_METER =
+      new SdkMeter(
+          PROVIDER_SHARED_STATE, SCOPE, Collections.emptyList(), MeterConfig.defaultConfig());
 
   @Test
   void stringRepresentation() {
     InstrumentBuilder builder =
         new InstrumentBuilder(
-                "instrument-name",
-                InstrumentType.COUNTER,
-                InstrumentValueType.LONG,
-                PROVIDER_SHARED_STATE,
-                METER_SHARED_STATE)
+                "instrument-name", InstrumentType.COUNTER, InstrumentValueType.LONG, SDK_METER)
             .setDescription("instrument-description")
             .setUnit("instrument-unit")
             .setAdviceBuilder(Advice.builder());
@@ -57,11 +53,7 @@ class InstrumentBuilderTest {
   void toStringHelper() {
     InstrumentBuilder builder =
         new InstrumentBuilder(
-                "instrument-name",
-                InstrumentType.HISTOGRAM,
-                InstrumentValueType.DOUBLE,
-                PROVIDER_SHARED_STATE,
-                METER_SHARED_STATE)
+                "instrument-name", InstrumentType.HISTOGRAM, InstrumentValueType.DOUBLE, SDK_METER)
             .setDescription("instrument-description")
             .setUnit("instrument-unit")
             .setAdviceBuilder(Advice.builder());
