+++ sdk/metrics/src/test/java/io/opentelemetry/sdk/metrics/SdkObservableInstrumentTest.java
@@ -6,18 +6,16 @@
 package io.opentelemetry.sdk.metrics;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.never;
-import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.verify;
 
 import io.github.netmikey.logunit.api.LogCapturer;
 import io.opentelemetry.internal.testing.slf4j.SuppressLogger;
 import io.opentelemetry.sdk.common.InstrumentationScopeInfo;
-import io.opentelemetry.sdk.metrics.internal.MeterConfig;
 import io.opentelemetry.sdk.metrics.internal.descriptor.Advice;
 import io.opentelemetry.sdk.metrics.internal.descriptor.InstrumentDescriptor;
 import io.opentelemetry.sdk.metrics.internal.state.CallbackRegistration;
-import io.opentelemetry.sdk.metrics.internal.state.MeterSharedState;
 import io.opentelemetry.sdk.metrics.internal.state.SdkObservableMeasurement;
 import java.util.Collections;
 import org.junit.jupiter.api.BeforeEach;
@@ -30,18 +28,13 @@ class SdkObservableInstrumentTest {
   @RegisterExtension
   LogCapturer logs = LogCapturer.create().captureForType(SdkObservableInstrument.class);
 
-  private MeterSharedState meterSharedState;
   private CallbackRegistration callbackRegistration;
+  private SdkMeter sdkMeter;
   private SdkObservableInstrument observableInstrument;
 
   @BeforeEach
   void setup() {
-    meterSharedState =
-        spy(
-            MeterSharedState.create(
-                InstrumentationScopeInfo.empty(),
-                Collections.emptyList(),
-                MeterConfig.defaultConfig()));
+    sdkMeter = mock(SdkMeter.class);
     callbackRegistration =
         CallbackRegistration.create(
             Collections.singletonList(
@@ -57,7 +50,7 @@ class SdkObservableInstrumentTest {
                     Collections.emptyList())),
             () -> {});
 
-    observableInstrument = new SdkObservableInstrument(meterSharedState, callbackRegistration);
+    observableInstrument = new SdkObservableInstrument(sdkMeter, callbackRegistration);
   }
 
   @Test
@@ -65,13 +58,13 @@ class SdkObservableInstrumentTest {
   void close() {
     // First call to close should trigger remove from meter shared state
     observableInstrument.close();
-    verify(meterSharedState).removeCallback(callbackRegistration);
+    verify(sdkMeter).removeCallback(callbackRegistration);
     logs.assertDoesNotContain("has called close() multiple times.");
 
     // Close a second time should not trigger remove from meter shared state
-    Mockito.reset(meterSharedState);
+    Mockito.reset(sdkMeter);
     observableInstrument.close();
-    verify(meterSharedState, never()).removeCallback(callbackRegistration);
+    verify(sdkMeter, never()).removeCallback(callbackRegistration);
     logs.assertContains("has called close() multiple times.");
   }
